requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
pandas==2.1.4
python-multipart==0.0.6
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.23.3
openai==2.16.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25
